# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ORIONIS AI Security Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /threats:
    get:
      operationId: listThreats
      tags: [threats]
      summary: List active threats
      parameters:
        - name: severity
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of threats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Threat"
    post:
      operationId: createThreat
      tags: [threats]
      summary: Create threat intelligence entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThreatInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Threat"
...
  /threats/{id}:
    get:
      operationId: getThreat
      tags: [threats]
      summary: Get threat by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Threat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Threat"
        "404":
          description: Not found
    patch:
      operationId: updateThreat
      tags: [threats]
      summary: Update threat status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThreatPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Threat"

  # ─── Incidents ───────────────────────────────────────────────
...
  /workflows:
    get:
      operationId: listWorkflows
      tags: [workflows]
      summary: List workflows
      responses:
        "200":
          description: List of workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
    post:
      operationId: createWorkflow
      tags: [workflows]
      summary: Create a workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{id}:
    get:
      operationId: getWorkflow
      tags: [workflows]
      summary: Get workflow by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
    patch:
      operationId: updateWorkflow
      tags: [workflows]
      summary: Update workflow
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
    delete:
      operationId: deleteWorkflow
      tags: [workflows]
      summary: Delete workflow
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /workflows/{id}/execute:
    post:
      operationId: executeWorkflow
      tags: [workflows]
      summary: Execute a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Execution started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowExecution"
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List reports
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a new report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  # ─── Audit Logs ──────────────────────────────────────────────
...
components:
  schemas:
...
    Threat:
      type: object
      required: [id, threatId, type, severity, status, createdAt]
      properties:
        id:
          type: string
        threatId:
          type: string
        type:
          type: string
        severity:
          type: string
        sourceIp:
          type: ["string", "null"]
        targetAssetId:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        indicators:
          type: ["object", "null"]
          additionalProperties: true
        mitreTechniques:
          type: array
          items: { type: string }
        score:
          type: ["number", "null"]
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
    ThreatInput:
      type: object
      required: [type, severity, description]
      properties:
        type:
          type: string
        severity:
          type: string
        sourceIp:
          type: string
        targetAssetId:
          type: string
        description:
          type: string
        indicators:
          type: object
          additionalProperties: true
        mitreTechniques:
          type: array
          items: { type: string }
        score:
          type: number
    ThreatPatch:
      type: object
      properties:
        status:
          type: string
        score:
          type: number
...
    Workflow:
      type: object
      required: [id, name, triggerType, status, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        triggerType:
          type: string
        triggerConfig:
          type: object
          additionalProperties: true
        steps:
          type: array
          items: { type: object, additionalProperties: true }
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    WorkflowInput:
      type: object
      required: [name, triggerType]
      properties:
        name:
          type: string
        description:
          type: string
        triggerType:
          type: string
        triggerConfig:
          type: object
          additionalProperties: true
        steps:
          type: array
          items: { type: object, additionalProperties: true }
    WorkflowPatch:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        triggerConfig:
          type: object
          additionalProperties: true
        steps:
          type: array
          items: { type: object, additionalProperties: true }
    WorkflowExecution:
      type: object
      required: [executionId, workflowId, status, startedAt]
      properties:
        executionId:
          type: string
        workflowId:
          type: string
        status:
          type: string
        currentStep:
          type: ["number", "null"]
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]

    # ─── Compliance ────────────────────────────────────────────
...
    Report:
      type: object
      required: [id, title, type, status, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
        generatedBy:
          type: ["string", "null"]
        data:
          type: ["object", "null"]
          additionalProperties: true
        filePath:
          type: ["string", "null"]
        format:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    ReportInput:
      type: object
      required: [title, type]
      properties:
        title:
          type: string
        type:
          type: string
        format:
          type: string

    # ─── Audit Logs ────────────────────────────────────────────
...